requirements.txt
absl-py==1.3.0
asn1crypto==1.3.0
bleach==1.5.0
cachetools==4.2.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
conda==4.8.2
conda-pack==0.6.0
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.11.0
dataclasses==0.6
fonttools==4.38.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
html5lib==0.9999999
idna==2.8
imageio==2.24.0
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
munch==2.5.0
numpy==1.21.6
oauthlib==3.2.2
packaging==22.0
pandas==1.3.5
patsy==0.5.3
Pillow==9.2.0
pretrainedmodels==0.7.4
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel_yaml==0.15.87
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
six==1.14.0
statsmodels==0.13.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.7.0+cu110
torchaudio==0.7.0
torchvision==0.8.1+cu110
tqdm==4.42.1
typing_extensions==4.4.0
urllib3==1.25.8
Werkzeug==2.2.2
zipp==3.10.0
